+++ internal/pkg/utils/ioutil/writer_test.go
@@ -1,4 +1,4 @@
-package utils
+package ioutil
 
 import (
 	"bufio"
@@ -8,17 +8,17 @@ import (
 	"go.uber.org/zap/buffer"
 )
 
-func TestNewBufferWriter(t *testing.T) {
+func TestBufferedWriter(t *testing.T) {
 	t.Parallel()
-	writer := NewBufferWriter()
+	writer := NewBufferedWriter()
 	_, err := writer.WriteString("test")
 	assert.NoError(t, err)
 	assert.Equal(t, "test", writer.String())
 }
 
-func TestConnectTo(t *testing.T) {
+func TestBufferedWriter_ConnectTo(t *testing.T) {
 	t.Parallel()
-	writer := NewBufferWriter()
+	writer := NewBufferedWriter()
 	otherBuffer := buffer.Buffer{}
 	otherWriter := bufio.NewWriter(&otherBuffer)
 	writer.ConnectTo(otherWriter)
